gateway/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
python-dotenv==1.0.1
websockets==12.0
